+++ packages/react-native-audio-api/common/cpp/utils/AudioUtils.cpp
@@ -1,11 +1,11 @@
 #include "AudioUtils.h"
 
 namespace audioapi::AudioUtils {
-size_t timeToSampleFrame(double time, int sampleRate) {
+size_t timeToSampleFrame(double time, float sampleRate) {
   return static_cast<size_t>(time * sampleRate);
 }
 
-double sampleFrameToTime(int sampleFrame, int sampleRate) {
+double sampleFrameToTime(int sampleFrame, float sampleRate) {
   return static_cast<double>(sampleFrame) / sampleRate;
 }
