+++ packages/react-native-audio-api/common/cpp/utils/AudioUtils.h
@@ -6,8 +6,8 @@
 
 namespace audioapi::AudioUtils {
 
-size_t timeToSampleFrame(double time, int sampleRate);
-double sampleFrameToTime(int sampleFrame, int sampleRate);
+size_t timeToSampleFrame(double time, float sampleRate);
+double sampleFrameToTime(int sampleFrame, float sampleRate);
 
 float linearInterpolate(const float *source, size_t firstIndex, size_t secondIndex, float factor);
